Moved Month stuff into its own module, for cleanliness.
